Separate out an auto-load module
